+++ src/onegov.ballot/HISTORY.rst
@@ -1,5 +1,7 @@
 Changelog
 ---------
+1.9.0 (2017-11-23)
+~~~~~~~~~~~~~~~~~~~
 
 - Adds title with translations to ballots.
   [msom]